Increase Meson warning level to 3 (#61)
* Add tag-triggered release workflow for automatic GitHub Releases (#56)

Pushing a vMAJOR.MINOR.PATCH tag now builds, tests and publishes a GitHub
Release with per-platform archives, replacing the manual build-and-upload
process that produced the single hand-made v2.7.0.apple asset.

.github/workflows/release.yml has three jobs:

- validate: rejects tags that are not vMAJOR.MINOR.PATCH with an optional
  pre-release suffix, and fails if the tag disagrees with meson.build's
  version: or src/version.h's AH_VERSION. Runs first so a bad tag costs
  seconds rather than three full builds.
- build: the same 3-platform matrix and dependency install as ci.yml. Gates
  on meson test at the default debug buildtype, deliberately matching CI,
  since two documented lifetime bugs in this codebase reproduce at -O0 but
  pass silently at -O2. Ships a separate --buildtype=release binary, stripped
  and (on macOS) re-signed, because stripping invalidates the linker's ad-hoc
  signature and an arm64 binary would otherwise die with "Killed: 9".
- publish: gated on all three builds, so a failing platform can never leave a
  partial release visible. Creates the release as a draft, attaches every
  asset plus a combined SHA256SUMS.txt, and publishes last. Re-runs reuse the
  existing release and replace assets in place, leaving hand-edited notes
  alone; a draft left by an interrupted run is published by the next one.

Archives are alh-<version>-<platform>.tar.gz containing the binary, LICENSE,
BUILD-INFO.txt, the docs and terrain_bitmaps. A tarball is used rather than a
raw binary so the executable bit survives download.

The binaries stay dynamically linked against wxWidgets, matching the existing
asset; self-contained bundling is packaging work the issue puts out of scope,
so the prerequisite is documented instead.

Docs: README.md gains a Releases section covering tag format, how to cut a
release, what gets published, runtime prerequisites including the macOS
Gatekeeper quarantine, and how to recover a failed run. doc/readme.html now
points at the GitHub Releases page instead of a dead geocities link, describes
unpacking the archive, and has its misplaced #contacts anchor moved onto the
heading it names. doc/readme_bin.txt is rewritten from a 2006 Mandriva note
into a description of the current archive.


* Increase Meson warning level to 3

Fix all level-3 warnings across ah and parser-tests and enforce werror in CI.

Closes #45


* Point shipped docs at the README on GitHub, not a relative path (#56)

doc/readme.html and doc/readme_bin.txt now ship inside the release archives,
where there is no repository root above doc/ - so "../README.md" and "see
README.md in the repository root" were dead ends for anyone reading them from
an unpacked archive. Use the canonical GitHub URL instead, which resolves both
from a checkout and from an archive.

This also fixes the pre-existing ../README.md link in the "Compiling the
sources" section, which was correct while readme.html only ever lived in a
checkout, and became reachable-but-broken once this branch started shipping it.


* Fix release workflow token permissions

Grant actions:write at workflow scope for artifact uploads and actions:read in publish job for artifact downloads.


* Fix linux werror sign-compare in unit pane

Cast wx list item data to long in the GetUnit assertion to keep Linux -Werror builds clean.


---------

+++ src/gui/unitfilterdlg.cpp
@@ -223,7 +223,7 @@ void CUnitFilterDlg::Init()
     int          count;
     int          i;
     const char * item;
-    int          skilllen;
+    size_t       skilllen;
     std::string         S;
 
     skilllen    = strlen(PRP_SKILL_POSTFIX);
@@ -239,13 +239,13 @@ void CUnitFilterDlg::Init()
             // do not show 'skill days' property
             S = item;
             if (!IsASkillRelatedProperty(S.c_str()) ||
-                FindSubStrR(S, PRP_SKILL_POSTFIX) == S.size()-skilllen)
+                static_cast<size_t>(FindSubStrR(S, PRP_SKILL_POSTFIX)) == S.size()-skilllen)
             {
                  m_cbProperty[count]->Append(wxString::FromAscii(item));
             }
         }
 
-        for (i=0; (unsigned)i<sizeof(UNIT_FILTER_OPERATION)/sizeof(*UNIT_FILTER_OPERATION); i++)
+        for (i=0; i<static_cast<int>(sizeof(UNIT_FILTER_OPERATION)/sizeof(*UNIT_FILTER_OPERATION)); i++)
         {
             item = UNIT_FILTER_OPERATION[i];
             m_cbCompare[count]->Append(wxString::FromAscii(item));
